Rename module optbase as optzbase for clarity
