+++ ams/opt/__init__.py
@@ -2,7 +2,7 @@
 Module for optimization modeling.
 """
 
-from ams.opt.optbase import OptzBase, ensure_symbols, ensure_mats_and_parsed  # NOQA
+from ams.opt.optzbase import OptzBase, ensure_symbols, ensure_mats_and_parsed  # NOQA
 from ams.opt.var import Var  # NOQA
 from ams.opt.exprcalc import ExpressionCalc  # NOQA
 from ams.opt.param import Param  # NOQA
